pyproject.toml
[tool.poetry]
name = "thebook"
version = "0.1.0"
description = "The Book - Hackerspace Management System"
authors = ["Renne Rocha <renne@rocha.dev.br>"]
license = "AGPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
python-decouple = "^3.8"
dj-database-url = "^2.2.0"
gunicorn = "^23.0.0"
psycopg = {extras = ["binary"], version = "^3.2.3"}
django-storages = {extras = ["s3"], version = "^1.14.4"}
ofxtools = "^0.9.5"
ofxparse = {git = "https://github.com/rennerocha/ofxparse.git", rev = "master"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-django = "^4.9.0"
taskipy = "^1.13.0"
pre-commit = "^3.8.0"
ipdb = "^0.13.13"
ipython = "^8.27.0"
pytest-cov = "^5.0.0"
model-bakery = "^1.19.5"
pytest-freezegun = "^0.4.2"
pytest-mock = "^3.14.0"
djade = "^1.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
collectstatic = "python manage.py collectstatic"
makemigrations = "python manage.py makemigrations"
migrate = "python manage.py migrate"
run = "python manage.py runserver"
test = "pytest -s --cov=thebook"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "thebook.settings"

[tool.isort]
known_django = ["django"]
known_first_party = ["thebook"]
known_third_party = ["django", "model_bakery", "pytest"]
profile = "black"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "DJANGO", "FIRSTPARTY", "LOCALFOLDER"]
